fix: fix results
